Allow users to specify a more granular log expire interval
Changed LogExpireInterval to be an arbitrary time interval instead
of just an integer number of days.  The time interval is specified
just like time interval literals in Bro scripts (except that negative
values are not allowed, and only time units "day", "hr", and "min" are
recognized).

For backward compatibility, an integer value without a time unit is
still interpreted as a number of days.