Eliminate unnecessary writes to the state db
Added a check to set_state() to prevent redundant writes to the state db.
However, this wasn't sufficient, because broctl was always writing the
config hash values to the state db upon startup, so the call to
read_state() was moved to happen earlier in the startup of broctl.

Also, when writing the standalone or cluster layout files, some
code was simplified to avoid unnecessarily changing the manager port
number.