Prevent start helper from getting in infinite loop
In the unlikely scenario that the run-bro script fails to create a
pid file (for example, if the working directory is owned by another
user), then the start helper script waits forever for the pid file
to be created.  In the meantime, Bro is started but broctl doesn't
have the PID and can't stop Bro.

Fixed by creating and removing a test file (and if that fails,
exit with a useful error message) right before launching
the run-bro script.