Change node hostname resolution to be more consistent
Each time broctl starts up, it tries to convert node host values (could
be either hostname or IP address) to node addr (always an IP address).
Since the tests use "host=localhost", sometimes localhost would
resolve to 127.0.0.1 and sometimes to ::1, and this difference would
cause broctl to assume that the node config was changed and it would
issue a warning message.  As a result, there were sporadic test failures.

To avoid this confusion, now broctl just uses the first IPv4 address in
the list, or the first address when no IPv4 address is found.