Change default value of the StatusCmdShowAll option
Changed the default value of the StatusCmdShowAll option so that
the broctl status command runs faster.  This also helps the tests
run faster (only one test actually needs the "Peers" column in
the status output).

Also changed the status-timefmt test so that it can be run in parallel
with the other tests.