Improve broctl behavior when unable to stop a node
Previously, when broctl failed to signal a Bro process to stop (i.e.,
the SIGTERM was not sent for any reason), it would try to send a SIGKILL
and it would loop many times checking on that node's status.  This took
a long time, and could output many redundant error messages, and often
didn't actually accomplish anything useful.  Fixed this so that it
outputs a more descriptive error message and then it will give up on
that node.