Fix race condition in reading/writing broctl-config.sh
Changed broctl to update the broctl-config.sh file as an atomic operation.
When creating broctl-config.sh, broctl will now write a tmp file first,
and then rename it to broctl-config.sh so that if another process (such
as archive-log, spawned from Bro) tries to read the file, the file
will never be empty or in a partially-written state.