Improve crash reports that have no backtrace
Changed the text of the crash reports that broctl mails so that if no
backtrace is included then the crash report will now explain what to do
in order to get a backtrace (there is a separate message depending on
whether there's no core file, no gdb, or both are missing).  Only when
a backtrace is included will the crash report tell the user to mail the
crash report to the bro team.

Also added gdb to the list of recommended prerequisites in the
documentation.

Updated the diag command tests and baselines.