Fix bug in setting log rotation interval for logger node
Fixed setting the log rotation interval when a logger node is used
(the value was always being set on the manager node).

Also set a new script boolean that the cluster framework scripts need
in order to determine if there is a logger node or not (previously, the
node name "logger" was hard-coded in the cluster framework scripts).

Also changed the tag name "logger" used in the "nodes" method to
"loggers" in order to be more consistent with how proxies and workers
are handled (and also in case we ever want to support multiple
loggers in a cluster).