Improve some broctl error messages
Improved the text of some error messages.  Also changed a few error
messages to warning messages.

Fixed the is_cfg_changed function to return False when it can't read
the config files, so that the user can continue using broctl in that
situation.

Removed some unneeded code in the make_layout and warn_broctl_install
functions.