Fix the plugin getGlobalOption() function
Fixed the getGlobalOption() function in the plugin API.  It did not
convert its argument to lowercase, and could return the value of a
state variable.

Also simplified some code by replacing the config has_attr() function
with a new function get_option(), which helps reduce the number of places
in the code where keys are converted to lowercase.