Fix the start-memlimit.test
The test would always fail on any system which didn't have a default
value of "unlimited" for the memory limits.  Fixed now so that it
just checks that the new value is not already set by default (because
we're just trying to test that the value can be changed, and we
don't care what the default value is).