Add error checking of archive-log timestamp parameters
Check if the format of the timestamp command-line parameters matches
the required format.  If not, exit with an error message.  This will
prevent archive-log from creating an archived log file with a corrupt
filename or in a directory with a corrupt name.

Also simplified the code that gets the current century.