Fix usage of run_localcmd()
Fixed several places in the code where an IndexError could potentially
be raised if no output is returned from run_localcmd().

Fixed a few other uses of run_localcmd() to check and handle the output
string correctly.

Also simplified the function to no longer take an optional argument
"donotcaptureoutput", and removed some debug output that didn't really
show useful info.  Also, instead of returning the output as a list of
strings, now it returns the output as a string.

These changes make the code slightly simpler and more robust.