Fix usage of run_cmds()
Fixed several places in the code where an IndexError could potentially
be raised if no output is returned from run_cmds(), and also fixed
a few potential ValueError cases if broctl gets unexpected output from
a command.

Fixed a few other uses of run_cmds() to check and handle the output
string correctly.

Also, instead of returning the output as a list of strings, now it returns
the output as a string.

Also fixed a few cases where the ps plugin didn't return non-zero when an
error occurred.