Fix the stats-to-csv script
Fixed a couple of bugs (proxies were being handled like workers, and
it was assuming that the manager is named "manager").  Also added
more error checking, and removed unused code.  Also changed the script
so that it ignores bad lines in stats.log and meta.dat (but still
prints an error message) instead of exiting.