Simplify some conditional assignments
Replaced code of the form "X = C and A or B" with "X = A if C else B"
because the latter is easier to understand and is more predictable
(e.g. if C evaluates to True and A to False, then X is assigned the
value of B in the first form, which is probably not what the programmer
intended).

Also simplified similar "if/else" statements that were also just
assigning to a single variable.