Fix sorting of node names
Fixed the sorting of node names.  Previously, a node such as "worker-10"
would be output after "worker-1" and before "worker-2".  Now the order
of names is based on the "count" node attribute instead of the name,
so the order of nodes (of a specific type) in node.cfg can affect the
order shown in the output.

Updated some test baselines.