Fix archive-log to handle logs that are already compressed
When logs are already compressed, then archive-log needs to get
the file extension preceding the ".gz" extension before running
the "make-archive-name" script (for example, for a log file named
"conn.2017-08-29-14-06-56.log.gz", we need to pass "conn.log" to
the "make-archive-name" script).

Also, changed the way archive-log decides whether or not to compress
a log.  It no longer checks the log writer type, since that information
by itself is not useful.  Instead, it now checks if a log has already
been gzipped.