Fix race conditions in a couple of tests
Fixed a race condition in the "update" command test.  It was in the
wrong serialization set, which was causing other serialized tests
to fail occasionally.

Fixed a race condition in the "start-slowstart" command test.  Sometimes
"broctl stop" would kill a process before the bro state file was
created, which resulted in the node being put in the "crashed" state.
Fixed the bro__test script by creating the state file first (the
"initializing" state), and setting the SIGTERM signal handler earlier
to ensure the state file is updated (the "terminated" state).
Also replaced the wait time with a synchronization mechanism to avoid
having to guess how long the wait time should be.  Also removed the
second "broctl status" check, as that wasn't really relevant to this
test.

Cleaned up the bro__test script (reduced global variables, add more
code comments, etc.).