Remove redundant output from broctl capstats command
The broctl capstats command would erroneously output a "total" line
when more than one node was specified on the command line, but
only one result was output (e.g., there is no output for nodes
such as manager that don't have a netif).  Added a test case
that would have caught this bug.

Also removed blank lines at the start and end of the output to be
more consistent with other broctl commands.