Reduce the size of the Bro build when running broctl tests
Added some additional configure options to reduce the size of the bro
build (on Travis CI, the bro binary went from 250MB to 10MB).  As these
options are extremely unlikely to be used in a previous bro build, they
are not applied when a user specifies a value for the env. variable
BROCTL_TEST_USEBUILD (note that a user can specify any additional options
via the BROCTL_TEST_BUILDARGS env. variable).

Also combined "make" and "make install" into one step to reduce the build
time slightly.