Update the stats-to-csv script for broker and loggers
Add support for logger nodes so that the script still works when run
on a logger node.

With the switch to broker, there are no longer any "child" entries in
the stats.log, so the script was not writing any data to the files in
the "logs/stats/www/" directory.  For now, keep support for "child" entries
so that the script works for both old and new stats.log files.