Update the broctl top command to not show the "Proc" column
Removed the "Proc" column (it showed "parent" or "child" for each bro
process) because after the switch to broker, Bro no longer runs as
two processes.

This simplified some code, and now the broctl "top" command should
run faster because it no longer needs to run the "get-childs" helper
script.

Note that these changes do not change the format of the broctl stats.log
file (the "parent" field should probably be removed in a future release).