Add WebSocket API support for interacting with nodes
This introduces a new configuration option called UseWebsocket.
If set, the manager node will spawn a WebSocket server and listen on
WebSocketHost (default 127.0.0.1) and WebSocketPort (default 27759/tcp).

Zeekctl will establish a short-lived WebSocket connection to the manager
and publish command events to node specific topics, then waiting for the
reply event from the corresponding node. That is, the manager just forwards
events between Zeekctl and other nodes. This works for both cluster backends,
Broker and ZeroMQ. When UseWebSocket is not set, there will be a warning
that print and netstats will not function.

This does not support TLS configuration for the WebSocket server. If you need
this, put a proxy (nginx, caddy, traefik, ...) in front of Zeek and use the
new WebSocketUrl setting of zeekctl.cfg to point it at the proxy with wss://.

This functionality requires the Python package websockets with at least
version 11.0. If that's not found and UseWebSocket is set, zeekctl will
warn about the issue during startup.