Avoid rebuilding Zeek for each test run
The test `Makefile` would previously always run the `buildzeek` target
when running the default target. This is not necessary as building Zeek
once is sufficient.

In this patch we also fix the setup so it works with parallel builds.
Before it might have run cleanup, building of Zeek and tests in parallel
which makes no sense.