cluster_backend_zeromq: Allocate XPUB/XSUB for proxy using ZeekPort
This modifies the cluster_backend_zeromq plugin to read the ZeekPort
option write back ZeekPort + 2, using up two ports for ZeroMQ.

This causes a bunch of baseline diffs because the generated cluster
layouts start a different ports now. This is a bit annoying because
technically there's no reason for manager, worker and proxy nodes to
have the $p member in the cluster-layout. At the same time, keep it
there to make it easier to quickly switch back to Broker for testing
without needing to re-generate the cluster-layout.

We could have the cluster backend plugin propagate to install.py
information about which nodes require port allocations, but I'm a
bit conflicted if that is worth it (and see the argument for having
ports populated for testing Broker above).